project/dashboards/supplychain-dashboard/pyproject.toml
[project]
name = "supplychain-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "debugpy>=1.8.14",
    "dj-database-url>=2.3.0",
    "django>=4.2",
    "django-cors-headers>=3.14",
    "django-filter>=25.1",
    "django-guardian<3",
    "django-stubs>=5.2.0",
    "djangorestframework>=3.14,<4",
    "djangorestframework-api-key>=3.1.0",
    "djangorestframework-simplejwt>=5.2",
    "gunicorn>=20.1",
    "iota-sdk>=1.1.4",
    "psycopg2-binary>=2.9",
    "python-dotenv>=0.21",
    "qrcode>=8.2",
    "requests>=2.32.3",
]
